requirements.txt
ultralytics
gradio
opencv-python
pillow
numpy
roboflow
pyyaml
